autorun configure with first run
